Code refactoring+ notebook updates
Code refactoring in order to be more consistent with basic programming recommendations. The importation functions are now isolated in a dedicated module named import_fc.
additionnaly, some updates of documentation to help unexperienced pip/python/jupyter users install and run Karstnet.
