+++ README.md
@@ -13,13 +13,17 @@ Version 0.1.0 - December 2018
 
 ## Installation
 
+Download the Karstnet folder. 
+
 Install from source (from project main directory):
 `pip install .`
 
 If you want to run it directly from source (useful for development):
 ` pip install -e .`
 
-
+Both these options can also be run without moving previously in the karstnet folder. 
+In that case, just type in your Anaconda prompt :
+` pip install -e your\path\to\karstnet` 
 
 ## Testing
 
@@ -27,10 +31,31 @@ From source directory, and after instaling **karstnet** run:
 
 `pytest tests/test_karstnet.py`
 
+## In Jupyter notebooks
 
+Example of jupyter notebooks are provided to help you use **karstnet**. 
+To use **karstnet** in notebooks, you just have to write
+
+`import karstnet as kn`
+
+A call-test function is available to help you check if the package i ready to use : just type: 
+`kn.test_kn()`
 
 ## Documentation
 
 The html documentation is available in the sub directory:  ``docs/_build/html/index.html``
 
 Also available online at: https://karstnet.readthedocs.io/
+
+## Reference
+
+The karstnet package implements the statistical metrics that were
+investigated and discussed in:
+Collon, P., Bernasconi D., Vuilleumier C., and Renard P., 2017, Statistical
+metrics for the characterization of karst network geometry and topology.
+Geomorphology. 283: 122-142 `doi:10.1016/j.geomorph.2017.01.034
+<http://dx.doi.org/doi:10.1016/j.geomorph.2017.01.034>`_
+
+The paper can be downloaded `here
+<http://members.unine.ch/philippe.renard/articles/collon2017.pdf>` or `here <https://hal.univ-lorraine.fr/hal-01468055v1>`_. 
+
