+++ karstnet/__init__.py
@@ -1,1 +1,2 @@
 from karstnet.base import *
+from karstnet.import_fc import *
